🔨 Refactor(#112): style -> styles 파일명으로 모두 변경
